+++ src/index.ts
@@ -5,6 +5,7 @@ import { AvatarEndpoint } from "./endpoints/avatar";
 import { ComponentEndpoint } from "./endpoints/component";
 import { ConfigurationEndpoint } from "./endpoints/configuration";
 import { CustomFieldEndpoint } from "./endpoints/customField";
+import { DashboardEndpoint } from "./endpoints/dashboard";
 import { PermissionsEndpoint } from "./endpoints/permissions";
 import { Basic, BasicAuth } from "./types";
 
@@ -68,9 +69,14 @@ export class JiraServerConnector {
     configuration: ConfigurationEndpoint;
 
     /**
-     * Contains all operations from '/rest/api/latest/customFields' and '/rest/api/latest/customFieldOptions''.
+     * Contains all operations from '/rest/api/latest/customFields' and '/rest/api/latest/customFieldOptions'.
      */
     customFields: CustomFieldEndpoint;
+
+    /**
+     * Contains all operations from '/rest/api/latest/dashboard'.
+     */
+    dashboards: DashboardEndpoint;
 
     /**
      * Instance new Jira Server Connector with user credentials and Jira host.
@@ -86,6 +92,7 @@ export class JiraServerConnector {
         this.component = new ComponentEndpoint(this.auth);
         this.configuration = new ConfigurationEndpoint(this.auth);
         this.customFields = new CustomFieldEndpoint(this.auth);
+        this.dashboards = new DashboardEndpoint(this.auth);
     }